Добавлен интерфейс управления временем жизни. Если время жизни == 0, объект не генинирируется

+++ src/factory/AbstractFactory.java
@@ -7,13 +7,13 @@ import model.Truck;
 public interface AbstractFactory {
     PassengerCar createPassengerCar();
 
-    PassengerCar createPassengerCar(float x, float y, float birthTime); // х и у - координаты
+    PassengerCar createPassengerCar(float x, float y, long birthTime); // х и у - координаты
 
-    PassengerCar createPassengerCar(float x, float y, float birthTime, float lifetime); // х и у - координаты
+    PassengerCar createPassengerCar(float x, float y, long birthTime, long lifetime); // х и у - координаты
 
     Truck createTruck();
 
-    Truck createTruck(float x, float y, float birthTime);
+    Truck createTruck(float x, float y, long birthTime);
 
-    Truck createTruck(float x, float y,  float birthTime, float lifetime);
+    Truck createTruck(float x, float y,  long birthTime, long lifetime);
 }
